[remove] auth2 관련 객체 삭제 #14 - JWT 기반 인증 서비스 구현 후 추후 기능 추가
