--- src/main/kotlin/dev/codeswamp/global/auth/config/SecurityConfig.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.config
-
-class SecurityConfig {
-}
