--- src/main/kotlin/dev/codeswamp/global/auth/controller/RefreshTokenController.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.controller
-
-class RefreshTokenController {
-}
