--- src/main/kotlin/dev/codeswamp/global/auth/entity/CustomUserDetails.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.entity
-
-class CustomUserDetails {
-}
