--- src/main/kotlin/dev/codeswamp/global/auth/entity/oauth2/CustomOAuth2User.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.entity.oauth2
-
-class CustomOAuth2User {
-}
