--- src/main/kotlin/dev/codeswamp/global/auth/entity/oauth2/OAuth2GoogleUserInfo.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.entity.oauth2
-
-class OAuth2GoogleUserInfo {
-}
