--- src/main/kotlin/dev/codeswamp/global/auth/filter/JwtAuthenticationFilter.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.filter
-
-class JwtAuthenticationFilter {
-}
