--- src/main/kotlin/dev/codeswamp/global/auth/handler/OAuth2FailureHandler.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.handler
-
-class OAuth2FailureHandler {
-}
