+++ src/main/kotlin/dev/codeswamp/global/auth/infrastructure/config/SecurityConfig.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.global.auth.infrastructure.config
+
+class SecurityConfig {
+}
