+++ src/main/kotlin/dev/codeswamp/global/auth/infrastructure/jwt/CustomUserDetails.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.global.auth.infrastructure.jwt
+
+class CustomUserDetails {
+}
