+++ src/main/kotlin/dev/codeswamp/global/auth/infrastructure/jwt/filter/JwtAuthenticationFilter.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.global.auth.infrastructure.jwt.filter
+
+class JwtAuthenticationFilter {
+}
