+++ src/main/kotlin/dev/codeswamp/global/auth/infrastructure/jwt/service/JwtService.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.global.auth.infrastructure.jwt.service
+
+interface JwtService {
+}
