+++ src/main/kotlin/dev/codeswamp/global/auth/infrastructure/jwt/service/JwtServiceImpl.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.global.auth.infrastructure.jwt.service
+
+class JwtServiceImpl {
+}
