+++ src/main/kotlin/dev/codeswamp/global/auth/presentation/controller/RefreshTokenController.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.global.auth.presentation.controller
+
+class RefreshTokenController {
+}
