--- src/main/kotlin/dev/codeswamp/global/auth/service/CustomOAuth2Service.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.service
-
-class CustomOAuth2Service {
-}
