--- src/main/kotlin/dev/codeswamp/global/auth/service/JwtService.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.service
-
-interface JwtService {
-}
