--- src/main/kotlin/dev/codeswamp/global/auth/service/JwtServiceImpl.kt
@@ -1,4 +1,0 @@
-package dev.codeswamp.global.auth.service
-
-class JwtServiceImpl {
-}
